Implementar mapeos JPA, patrón DAO/Singleton, menú interactivo con InputUtil y manejo de excepciones (Java 21)

+++ src/main/java/ar/edu/unju/escmi/dao/FacturaDAO.java
@@ -1,11 +1,10 @@
 package ar.edu.unju.escmi.dao;
 
 import ar.edu.unju.escmi.model.Factura;
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface FacturaDAO extends GenericDAO<Factura, Long> {
     List<Factura> findActive();
 
-    List<Factura> findByTotalGreaterThan(BigDecimal amount);
+    List<Factura> findByTotalGreaterThan(double amount);
 }
